Trakt api migration (#46)
* Update imports

* Inject tmdb api key and update imports.

* Fix status bar display when TopAppBar is displayed.

* Rename function name.

* Add trakt auth implementation.

* Move trakt logic code to android core module since this is Android related.

* Add URL to HttpClient. This fixes the an issue where logged url is localhost.

* Add trakt user table & update imports

* Minor cleanup, update show order param.

* Update dependencies.

* Add trakt network & cache logic.

* Update composable components.

* Provide strings.

* Remove dependency block.

* Load keys on debug build & update dependencies.

* Add trakt dependencies

* Add trakt modules

* Update used libraries & update API key setup instructions.

* Rename more to settings & update icon.

* Implement trakt auth sign in.

* Comment out gradle caching. There's an issue with kmm support. Remove deprecated configs.

* Add appauth & activity dependencies.

* Update show instead of replacing the whole object

* Add encodeDefaults allowing us to use post.

* Implement logic to create trakt list and add shows to the list.

* Add button with loading indicator

* Update padding.

* Add loading indicator when following/unfollowing show & some minor cleanup

* Rename repository class.

* Update show queries and make get show query nullable.

* Fetch & sync trakt followed shows.

* Add flow param.

* Add manifestPlaceholders

* Invoke sync workManager

* Add function to sync followed shows when device is idle.

* Cleanup follow show implementation & sync with trakt.

* Allow non-logged in users to cache locally.

* Create worker to sync TMDB art work.

* Update ID type to Int from Long

* Add trakt API calls

* Update tables, rename columns and update types since we've made trakt our main source of data.

* Update adapter type from Int to String

* Add double formatter & rename class.

* Use trakt endpoint to fetch data & do some cleanup.

* Move database module class to app module.

* Add combine function.

* Add compose dependency.

* Upgrade kotlin version to 1.7.10 & bump up dependency versions

* House tidying.

- Remove unused functions.
- Update show queries.
- Rename some classes.

* Update test data.

* Add trakt env keys

* Update network image composable after updating coil.

* Clean up discover implementation.

* Clean up db queries

* Add manifestPlaceholders

* Update dependencies.

* Rename module to shows

* Update imports.

* Enable buildfeatures config

* Delete discover module.

- Related classes have been moved to show module.

* Cleanup show repository implementation.

* Use trakt to fetch season information.

* Minor cleanup

* Update imports

* Use collectAsStateWithLifecycle and method references in compose screens

* Delete genre module.

* Delete genre module.

* Remove image urls from show

* Create image table and add cache implementation.

* Cleanup trakt & tmdb implementation.

* Cleanup trakt & tmdb image loading in shared module.

* Update poster component UI.

* Remove sync images worker.

* Remove sync images worker.

* Minor UI cleanup.

* Update dependency versions.

* Update todo tasks.

* Cleanup following implementation:

- Sort items by date added.
- Remove custom list creation implementation.
- Add watchlist endpoint

* Create profile module.

* Add profile implementation

* Update episode image loading.

* Exclude unused annotations

* Remove aired episodes UI.

* Update build gradle task

* Delete some tests (I promise, I will rewrite.)

* clean-up local properties access.

+++ shared/domain/last-air-episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/lastairepisodes/api/LastAirEpisodeRepository.kt
@@ -4,5 +4,5 @@ import com.thomaskioko.tvmaniac.core.db.AirEpisodesByShowId
 import kotlinx.coroutines.flow.Flow
 
 interface LastAirEpisodeRepository {
-    fun observeAirEpisodes(tvShowId: Long): Flow<List<AirEpisodesByShowId>>
+    fun observeAirEpisodes(tvShowId: Int): Flow<List<AirEpisodesByShowId>>
 }
